Long commit changing the way that Pans interact with the LRU cache
In order to keep old connections from hogging goroutines and memory
resources. PanCtl structs are now stored in the primary LRU cache and
contain a pointer to the Pan struct, PacketManifest channel, and a stop
channel to kill the goroutine.

PanRemover can be sent PanCtl structs for termination which handles
removal from LRU and closing the stop channel on the Pan struct which
frees the goroutine.